+++ src/main/java/com/cburch/logisim/std/arith/MultiplierHdlGeneratorFactory.java
@@ -17,6 +17,7 @@ import com.cburch.logisim.fpga.hdlgenerator.Hdl;
 import com.cburch.logisim.fpga.hdlgenerator.HdlParameters;
 import com.cburch.logisim.instance.Port;
 import com.cburch.logisim.util.LineBuffer;
+import java.util.HashMap;
 import java.util.Map;
 
 public class MultiplierHdlGeneratorFactory extends AbstractHdlGeneratorFactory {
@@ -27,20 +28,20 @@ public class MultiplierHdlGeneratorFactory extends AbstractHdlGeneratorFactory {
   private static final int CALC_BITS_ID = -2;
   private static final String UNSIGNED_STRING = "unsignedMultiplier";
   private static final int UNSIGNED_ID = -3;
-  private static final Map<AttributeOption, Integer> UNSIGNED_MAP =
-      Map.of(Comparator.UNSIGNED_OPTION, 1, Comparator.SIGNED_OPTION, 0);
+
+  public static final Map<AttributeOption, Integer> SIGNED_MAP = new HashMap<>() {{
+      put(Multiplier.UNSIGNED_OPTION, 0);
+      put(Multiplier.SIGNED_UNSIGNED_OPTION, 1);
+      put(Multiplier.SIGNED_OPTION, 2);
+    }};
+
 
   public MultiplierHdlGeneratorFactory() {
     super();
     myParametersList
         .add(NR_OF_BITS_STRING, NR_OF_BITS_ID)
         .add(CALC_BITS_STRING, CALC_BITS_ID, HdlParameters.MAP_MULTIPLY, 2)
-        .add(
-            UNSIGNED_STRING,
-            UNSIGNED_ID,
-            HdlParameters.MAP_ATTRIBUTE_OPTION,
-            Comparator.MODE_ATTR,
-            UNSIGNED_MAP);
+        .add(UNSIGNED_STRING, UNSIGNED_ID, HdlParameters.MAP_ATTRIBUTE_OPTION, Multiplier.MODE_ATTR, SIGNED_MAP);
     myWires
         .addWire("s_multResult", CALC_BITS_ID)
         .addWire("s_extendedcarryIn", CALC_BITS_ID)
@@ -61,13 +62,17 @@ public class MultiplierHdlGeneratorFactory extends AbstractHdlGeneratorFactory {
         .pair("calcBits", CALC_BITS_STRING);
     if (Hdl.isVhdl()) {
       contents.empty().addVhdlKeywords().add("""
-          s_multResult <= std_logic_vector(unsigned(inputA) * unsigned(inputB))
-                              {{when}} {{unsigned}} = 1 {{else}}
-                           std_logic_vector(signed(inputA) * signed(inputB));
-          s_extendedcarryIn({{calcBits}}-1 {{downto}} {{nrOfBits}}) <= ({{others}} => '0') {{when}} {{unsigned}} = 1 {{else}} ({{others}} => carryIn({{nrOfBits}}-1));
+          s_multResult <= std_logic_vector(unsigned(inputA)*unsigned(inputB))
+                              {{when}} {{unsigned}}= 1 {{else}}
+                           std_logic_vector(signed(inputA)*unsigned(inputB));
+                              {{when}} {{unsigned}}= 2 {{else}}
+                           std_logic_vector(signed(inputA)*signed(inputB));
+          s_extendedcarryIn({{calcBits}}-1 {{downto}} {{nrOfBits}}) <= ({{others}} => '0') {{when}} {{unsigned}} = 1 or {{unsigned}} = 2 {{else}} ({{others}} => carryIn({{nrOfBits}}-1));
           s_extendedcarryIn({{nrOfBits}}-1 {{downto}} 0) <= carryIn;
           s_newResult  <= std_logic_vector(unsigned(s_multResult) + unsigned(s_extendedcarryIn))
-                              {{when}} {{unsigned}} = 1 {{else}}
+                              {{when}} {{unsigned}}= 1 {{else}}
+                           std_logic_vector(signed(s_multResult) + unsigned(s_extendedcarryIn));
+                              {{when}} {{unsigned}}= 2 {{else}}
                            std_logic_vector(signed(s_multResult) + signed(s_extendedcarryIn));
           multHigh     <= s_newResult({{calcBits}}-1 {{downto}} {{nrOfBits}});
           multLow      <= s_newResult({{nrOfBits}}-1 {{downto}} 0);
@@ -88,6 +93,12 @@ public class MultiplierHdlGeneratorFactory extends AbstractHdlGeneratorFactory {
                        s_carryIn[{{calcBits}}-1:{{nrOfBits}}] = 0;
                        s_multUnsigned = $unsigned(inputA) * $unsigned(inputB);
                        s_intermediateResult = $unsigned(s_multUnsigned) + $unsigned(s_carryIn);
+                     end
+                  else if ({{unsigned}}== 2)
+                     begin
+                       s_carryIn[{{calcBits}}-1:{{nrOfBits}}] = 0;
+                       s_multUnsigned = $signed(inputA) * $unsigned(inputB);
+                       s_intermediateResult = $signed(s_multUnsigned) + $signed(s_carryIn);
                      end
                   else
                     begin